drop mki
