+++ labex.link.txt
@@ -1,3 +1,3 @@
 packages/dat-sdk-webpack
 packages/jupyterlab-dat
-packages/jupyterlab-dat-notebook-mk-ii
+packages/jupyterlab-dat-notebook
