+++ packages/_meta/src/index.ts
@@ -1,4 +1,3 @@
 import '@deathbeds/jupyterlab-dat';
-import '@deathbeds/jupyterlab-dat-notebook-mk-i';
-import '@deathbeds/jupyterlab-dat-notebook-mk-ii';
+import '@deathbeds/jupyterlab-dat-notebook';
 import '@deathbeds/dat-sdk-webpack';
