--- packages/jupyterlab-dat-notebook-mk-i/style/index.css
@@ -1,36 +1,0 @@
-.jp-Dat-Widget.jp-dat-mki {
-  padding: var(--jp-notebook-padding);
-  background-color: var(--jp-layout-color1);
-  display: flex;
-  flex-direction: column;
-  overflow-y: auto;
-}
-
-.jp-Dat-Widget.jp-dat-mki .jp-Dat-Widget-Main {
-  display: flex;
-  flex-direction: row;
-  flex: 1;
-  flex-wrap: wrap;
-}
-
-.jp-Dat-Widget.jp-dat-mki section {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: var(--jp-notebook-padding);
-  min-width: 20rem;
-}
-
-.jp-Dat-Widget.jp-dat-mki header {
-  display: flex;
-  flex-direction: row;
-}
-
-.jp-Dat-Widget.jp-dat-mki input {
-  font-family: var(--jp-code-font-family);
-  text-overflow: ellipsis;
-  width: 100%;
-  margin: var(--jp-notebook-padding);
-  text-align: center;
-}
