--- packages/jupyterlab-dat-notebook-mk-ii/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@deathbeds/jupyterlab-dat-notebook-mk-ii",
-  "version": "0.1.0",
-  "description": "a hopeful dat button for JupyterLab Notebooks",
-  "keywords": [
-    "jupyter",
-    "jupyterlab",
-    "jupyterlab-extension",
-    "dat",
-    "p2p"
-  ],
-  "homepage": "https://github.com/deathbeds/jupyterlab-dat",
-  "bugs": {
-    "url": "https://github.com/deathbeds/jupyterlab-dat/issues"
-  },
-  "license": "BSD-3-Clause",
-  "author": "dead pixels collective",
-  "files": [
-    "lib/**/*.{d.ts,eot,gif,html,jpg,js,js.map,json,png,svg,woff2,ttf}",
-    "style/**/*.{css,eot,gif,html,jpg,json,png,svg,woff2,ttf}"
-  ],
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "style": "style/index.css",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/deathbeds/jupyterlab-dat.git"
-  },
-  "scripts": {
-    "clean": "rimraf lib"
-  },
-  "peerDependencies": {
-    "@deathbeds/jupyterlab-dat": "0.2",
-    "@jupyterlab/notebook": "1",
-    "@jupyterlab/application": "1"
-  },
-  "devDependencies": {
-    "@deathbeds/jupyterlab-dat": "0.2",
-    "@jupyterlab/notebook": "1",
-    "@jupyterlab/application": "1"
-  },
-  "sideEffects": [
-    "style/*.css"
-  ],
-  "jupyterlab": {
-    "extension": "lib/plugin.js"
-  }
-}
