--- packages/jupyterlab-dat-notebook-mk-ii/src/index.ts
@@ -1,1 +1,0 @@
-export { CSS } from '@deathbeds/jupyterlab-dat';
