--- packages/jupyterlab-dat-notebook-mk-ii/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../tsconfigbase",
-  "compilerOptions": {
-    "outDir": "lib",
-    "rootDir": "src",
-    "tsBuildInfoFile": "lib/.tsbuildinfo"
-  },
-  "include": ["src/**/*"],
-  "references": [
-    {
-      "path": "../jupyterlab-dat"
-    }
-  ]
-}
