+++ packages/jupyterlab-dat-notebook/style/index.css
@@ -1,4 +1,4 @@
-.jp-Dat-Widget.jp-dat-mkii {
+.jp-Dat-Widget {
   padding: var(--jp-notebook-padding);
   background-color: var(--jp-layout-color1);
   display: flex;
@@ -6,14 +6,14 @@
   overflow-y: auto;
 }
 
-.jp-Dat-Widget.jp-dat-mkii .jp-Dat-Widget-Main {
+.jp-Dat-Widget .jp-Dat-Widget-Main {
   display: flex;
   flex-direction: row;
   flex: 1;
   flex-wrap: wrap;
 }
 
-.jp-Dat-Widget.jp-dat-mkii section {
+.jp-Dat-Widget section {
   flex: 1;
   display: flex;
   flex-direction: column;
@@ -22,30 +22,30 @@
   min-width: 20rem;
 }
 
-.jp-Dat-Widget.jp-dat-mkii header {
+.jp-Dat-Widget header {
   display: flex;
   flex-direction: row;
 }
 
-.jp-Dat-Widget.jp-dat-mkii input {
+.jp-Dat-Widget input {
   width: 100%;
   margin: 0 0 var(--jp-notebook-padding);
 }
 
-.jp-Dat-Widget.jp-dat-mkii details {
+.jp-Dat-Widget details {
   width: 100%;
 }
-.jp-Dat-Widget.jp-dat-mkii details label {
+.jp-Dat-Widget details label {
   display: flex;
   align-items: baseline;
 }
-.jp-Dat-Widget.jp-dat-mkii details label i {
+.jp-Dat-Widget details label i {
   flex: 0;
   min-width: 7em;
   text-align: right;
 }
-.jp-Dat-Widget.jp-dat-mkii details label input,
-.jp-Dat-Widget.jp-dat-mkii details label textarea {
+.jp-Dat-Widget details label input,
+.jp-Dat-Widget details label textarea {
   flex: 1;
   margin: 0 0 var(--jp-notebook-padding) var(--jp-notebook-padding);
 }
